03_my_implementations/chapter4_serving_patterns/05_prep_pred_pattern/resnet50_onnx_runtime/requirements.txt
torch==2.5.1
torchvision==0.20.1
onnxruntime==1.20.1
onnx==1.17.0
click==8.1.8
joblib==1.4.2
Pillow==11.0.0
numpy==2.2.0
scikit-learn==1.6.1
